refactor: StateCounts를 /type/study 로 분리
