+++ frontend/src/types/study.ts
@@ -1,0 +1,6 @@
+export interface StateCounts {
+  reviewCounts: number;
+  learningCounts: number;
+  overdueCounts: number;
+  newCounts: number;
+}
